rename files, expose proper api functions
